# Conflicts:
#	surf-roleplay-mechanic/src/main/kotlin/dev/slne/surf/roleplay/mechanic/MechanicRegistryImpl.kt

+++ surf-roleplay-mechanic/src/main/kotlin/dev/slne/surf/roleplay/mechanic/mechanics/license/player/LicensePlayerImpl.kt
@@ -1,5 +1,6 @@
 package dev.slne.surf.roleplay.mechanic.mechanics.license.player
 
+import com.github.shynixn.mccoroutine.folia.globalRegionDispatcher
 import dev.slne.surf.roleplay.api.mechanic.license.License
 import dev.slne.surf.roleplay.api.mechanic.license.PlayerLicense
 import dev.slne.surf.roleplay.api.mechanic.license.event.PlayerLicenseAddedEvent
@@ -9,9 +10,11 @@ import dev.slne.surf.roleplay.api.mechanic.license.utils.LicenseRemovedReason
 import dev.slne.surf.roleplay.api.mechanic.license.utils.UnobtainableReason
 import dev.slne.surf.roleplay.api.player.RpPlayer
 import dev.slne.surf.roleplay.mechanic.mechanics.license.LicenseService
+import dev.slne.surf.roleplay.mechanic.plugin
 import dev.slne.surf.surfapi.bukkit.api.extensions.server
 import dev.slne.surf.surfapi.core.api.util.freeze
 import dev.slne.surf.surfapi.core.api.util.mutableObjectSetOf
+import kotlinx.coroutines.withContext
 
 class LicensePlayerImpl(val rpPlayer: RpPlayer) : LicensePlayer {
 
@@ -25,30 +28,43 @@ class LicensePlayerImpl(val rpPlayer: RpPlayer) : LicensePlayer {
             return false to reason
         }
 
-        _licenses.add(LicenseService.createPlayerLicense(license, rpPlayer))
-
-        server.pluginManager.callEvent(
-            PlayerLicenseAddedEvent(
-                player = rpPlayer,
-                license = license,
-            )
+        val event = PlayerLicenseAddedEvent(
+            player = rpPlayer,
+            license = license,
         )
+
+        withContext(plugin.globalRegionDispatcher) {
+            server.pluginManager.callEvent(event)
+        }
+
+        if (event.isCancelled) {
+            return false to UnobtainableReason.EventCancelled(event.cancelReason)
+        }
+
+        _licenses.add(LicenseService.createPlayerLicense(license, rpPlayer))
 
         return true to null
     }
 
     override suspend fun removeLicense(license: License, reason: LicenseRemovedReason): Boolean {
+        val event = PlayerLicenseRemovedEvent(
+            player = rpPlayer,
+            license = license,
+            reason = reason
+        )
+
+        withContext(plugin.globalRegionDispatcher) {
+            server.pluginManager.callEvent(event)
+        }
+        
+        if (event.isCancelled) {
+            return false
+        }
+
         val result = LicenseService.removePlayerLicense(rpPlayer, license)
 
         if (result) {
             _licenses.removeIf { it.license == license }
-            server.pluginManager.callEvent(
-                PlayerLicenseRemovedEvent(
-                    player = rpPlayer,
-                    license = license,
-                    reason = reason
-                )
-            )
         }
 
         return result
